[hotfix] [tests] Speed up TaskManagerServicesTest
The TaskManagerServicesTest was extremely slow because it uses the PowerMockRunner
for a single test case. Additionally some other tests test different settings in
a loop which was quite slow because the PowerMockRunner instrumented all method calls.
Moving the test which required the PowerMockRunner into a separate test class sped up
things by a factor of 50.
